feat: add optimization example
